+++ src/app/App.tsx
@@ -1,10 +1,13 @@
 import PostManagerPage from "@pages/PostsManagerPage"
+import Layout from "@widgets/ui/Layout"
 import { BrowserRouter as Router } from "react-router-dom"
 
 const App = () => {
   return (
     <Router>
-      <PostManagerPage />
+      <Layout>
+        <PostManagerPage />
+      </Layout>
     </Router>
   )
 }
