+++ src/shared/ui/Button/Button.tsx
@@ -1,0 +1,13 @@
+import { forwardRef } from "react"
+import { buttonVariants } from "./button.css"
+import { VariantProps } from "class-variance-authority"
+
+interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof buttonVariants> {
+  className?: string
+}
+
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(({ className, variant, size, ...props }, ref) => {
+  return <button className={buttonVariants({ variant, size, className })} ref={ref} {...props} />
+})
+
+Button.displayName = "Button"
