+++ src/shared/ui/Card/Card.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+import { CardProps } from "./types"
+
+export const Card = forwardRef<HTMLDivElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+
+Card.displayName = "Card"
