+++ src/shared/ui/Card/CardHeader.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+import { CardProps } from "./types"
+
+export const CardHeader = forwardRef<HTMLDivElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+))
+
+CardHeader.displayName = "CardHeader"
