+++ src/shared/ui/Card/types.ts
@@ -1,0 +1,5 @@
+import { HTMLAttributes } from "react"
+
+export interface CardProps extends HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
