+++ src/shared/ui/Dialog/Dialog.tsx
@@ -1,0 +1,8 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+const Dialog = DialogPrimitive.Root
+const DialogTrigger = DialogPrimitive.Trigger
+const DialogPortal = DialogPrimitive.Portal
+const DialogOverlay = DialogPrimitive.Overlay
+
+export { Dialog, DialogTrigger, DialogPortal, DialogOverlay }
