+++ src/shared/ui/Dialog/index.ts
@@ -1,0 +1,4 @@
+export { Dialog, DialogTrigger, DialogPortal, DialogOverlay } from "./Dialog"
+export { DialogContent } from "./DialogContent"
+export { DialogHeader } from "./DialogHeader"
+export { DialogTitle } from "./DialogTitle"
