+++ src/shared/ui/Select/Select.tsx
@@ -1,0 +1,7 @@
+import * as SelectPrimitive from "@radix-ui/react-select"
+
+const Select = SelectPrimitive.Root
+const SelectGroup = SelectPrimitive.Group
+const SelectValue = SelectPrimitive.Value
+
+export { Select, SelectGroup, SelectValue }
