+++ src/shared/ui/Select/SelectTrigger.tsx
@@ -1,0 +1,17 @@
+import { forwardRef } from "react"
+import { ChevronDown } from "lucide-react"
+import * as SelectPrimitive from "@radix-ui/react-select"
+
+export const SelectTrigger = forwardRef<HTMLButtonElement, SelectPrimitive.SelectTriggerProps>(
+  ({ className, children, ...props }, ref) => (
+    <SelectPrimitive.Trigger
+      ref={ref}
+      className={`flex h-10 items-center justify-between rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      {...props}
+    >
+      {children}
+      <ChevronDown className="h-4 w-4 opacity-50" />
+    </SelectPrimitive.Trigger>
+  ),
+)
+SelectTrigger.displayName = SelectPrimitive.Trigger.displayName
