+++ src/shared/ui/Select/index.ts
@@ -1,0 +1,4 @@
+export { Select, SelectGroup, SelectValue } from "./Select"
+export { SelectContent } from "./SelectContent"
+export { SelectItem } from "./SelectItem"
+export { SelectTrigger } from "./SelectTrigger"
